+++ object/token_jwt.go
@@ -325,7 +325,7 @@ func refineUser(user *User) *User {
 
 func generateJwtToken(application *Application, user *User, nonce string, scope string, host string) (string, string, string, error) {
 	nowTime := time.Now()
-	expireTime := nowTime.Add(time.Duration(application.ExpireInHours) * time.Hour)
+	expireTime := nowTime.Add(time.Duration(application.ExpireInHours * float64(time.Hour)))
 	refreshExpireTime := nowTime.Add(time.Duration(application.RefreshExpireInHours) * time.Hour)
 	if application.RefreshExpireInHours == 0 {
 		refreshExpireTime = expireTime
